Bug 1649591 - Fix remote desktop graphics initialization. r=jrmuizel
If the device key is empty, we are most likely in a remote desktop environment. In this case we set the devicekey to an empty string so it can be handled later and not early out.

Differential Revision: https://phabricator.services.mozilla.com/D81782

UltraBlame original commit: a1db856139569cc49fe477a9e80b03c04e9c082c