Bug 1637744 - Create a new function for adding search engines, and a separate function for policy engines. r=daleharvey
The main aim here is to move the call to getEngineParams that currently happens before addEngineWithDetails. This is moved into addEngineWithDetails, so that it is next to where the search engine is actually created. This gets ready for the next step which will be to merge getEngineParams with the SearchEngine initWithMetadata and associated calls.

The side effects are that we need a specific function for policy engines to use, and that we now have only tests using addEngineWithDetails.

Differential Revision: https://phabricator.services.mozilla.com/D82348

UltraBlame original commit: 4a3af68e0ae18d5b94e3bc9e713f567a2d129b96