Bug 1651757 - Update frame rate stats dynamically on about:webrtc; r=ng
This adds an interval timer that periodically requests stats and updates
the values in frame rate table. The tables are given an id that matches
the report's pcid. This is necessary to disambiguate the tables if
multiple peerconnections are present.

This also changes the tables to use a monospace font. I think this is
easier to read anyway, and it reduces the amount of column movement when
tables are layed out again after the values change.

Differential Revision: https://phabricator.services.mozilla.com/D83089

UltraBlame original commit: 68b50e39d159a266a8f2fa52ba032119396596e4