Bug 1646378: Add WarpInlinedCall snapshot r=jandem
I based the InliningTree / CompileInfo code on IonBuilder::inlineScriptedCall.

In the future, if we want to inline cases where we didn't allocate a new ICScript, it should just be a matter of changing a few lines in maybeInlineIC. We can use the default ICScript off the target script.

(Note: we now check in maybeInlineCall that we don't already have a CallInlinedFunction. This should not normally happen, but could happen if we reset the warmup counter for a script.)

Differential Revision: https://phabricator.services.mozilla.com/D83184

UltraBlame original commit: ca2a620dc5a79760d5372b0996081d395805567b