Bug 1652675 - Define OPUS_ARM_MAY_HAVE_NEON{,_INTR} on non-android non-windows aarch64. r=kinetik
It somehow compiles fine currently without on Android, but relies on
implicit function declarations on Linux aarch64, which broke with
bug 822978, which turned the corresponding warning into an error.

It turns out OPUS_ARM_MAY* defines are set when the OPUS_ARM_PRESUME*
ones are set in the cmake build system upstream[1], and it also turns
out that not setting them actually disables some of the neon code. The
fixed point variant used for Android, however, fails to build in that
case because it relies on a function that is only available on arm
builds.

1. It's worth noting that the upstream autoconf build system does *not*
set OPUS_ARM_MAY* or OPUS_ARM_PRESUME* at all.

Differential Revision: https://phabricator.services.mozilla.com/D83474

UltraBlame original commit: 1d5986a1a6daa5b45d581d16652dbdde85ac4d86