Bug 1653094 - Update the android_wrench script to bypass the mozharness virtualenv. r=gbrown
android_wrench runs in the mozbuild virtualenv and doesn't need the mozharness
virtualenv. Trying to activate it fails because we don't do the setup for it,
so this patch just makes it so that attempting to activate it is a no-op.

The other change needed here is to make sure the is_emulator method works even
when called before the AndroidWrench constructor is done. This is needed because
it is used from query_abs_dirs and that now gets called earlier, before the
constructor is done.

Differential Revision: https://phabricator.services.mozilla.com/D83970

UltraBlame original commit: 94981e09a5d636e9172e517036184fd7bdfa8202