Bug 1652973 - [wdspec] Replace create_window fixture with new window command. r=jgraham
Now that each driver supports the new window command, we can make use of it
as replacement for the create_window fixture.

Given that there is no need to check the type of the new window,
this implementation doesn't add a check for that.

Differential Revision: https://phabricator.services.mozilla.com/D83800

UltraBlame original commit: 7ac192dc5d64a3e64d2a223053dba0cd6788157d