Bug 1652962 - Allow FinalizationQueueObject to outlive FinalizationRegistry object r=sfink
This rearranges things so that edge from FinalizationQueueObject to FinalizationRegistryObject is weak, and hence the latter can outlive the former. The weak edge from FinalizationRecordObject to FinalizationRegistryObject is replaced with a strong edge to FinalizationQueueObject.

This removes the requirement to trace anything during sweeping. The FinalizationQueueObject is kept alive by the strong edges from FinalizationRecordObjects.  As a result, they may live for an extra GC cycle more than they would otherwise (as the record objects did before), but this is not visible to users. This means we can remove GC support for doing extra marking work pushed during sweeping that was needed for this one case.

This also means we can remove the set of active records that needs to be maintained.

Differential Revision: https://phabricator.services.mozilla.com/D83760

UltraBlame original commit: 1f396812eae85fe830cc9f5c54f470e74eafe720