Bug 1578042 - Move mCompositionOpportunityId incrementing from EndTransaction to UpdateAndRender. r=nical
I got this wrong when I added non-WebRender support in the patch that added
the CompositionOpportunityId counter. I forgot that EndTransaction is not only
called when we composite, but also, who would have thought, when we process a
transaction.
We only want to increment mCompositionOpportunityId when we're actually about
to composite.

Differential Revision: https://phabricator.services.mozilla.com/D84111

UltraBlame original commit: e32e69b9bdc47dfb302a882710589be401a71e79