Bug 1653717: [taskgraph] Handle composite description nicely; r=Callek
When links to specific tasks in treeherder was added to the description in
Bug 1630710, This caused the descriptions for several derived tasks to include
the `{'task-reference': '...'}` wrapper in the derived description. This is
in addition to the already duplicate links to treeherder those tasks had.

To ameliorate this issue, keep track of the original description without the
treeherder link and use that in derived tasks.

Differential Revision: https://phabricator.services.mozilla.com/D84047

UltraBlame original commit: b116dcac12515cd09d25630185feb4f3caeafb5a